requirements.txt
langchain_core
python-dotenv
streamlit
langchain_community
langserve
fastapi
uvicorn
sse_starlette
bs4
faiss-cpu
pypdf
langchain-groq
faiss-cpu
pypdf
bs4
langchain-groq
groq
beautifulsoup4
cassio
wikipedia
langchainhub
sentence-transformers
datasets
langchain-huggingface
flask
langchain-google-genai
sentence-transformers
langchain-huggingface
langchain-google-genai
openai-whisper
ffmpeg
transformers 
soundfile 
duckduckgo-search
agno
google-genai
gspread 
oauth2client 
streamlit-mic-recorder 
openai
sacremoses
